refactor: community를 workout 도메인으로 병합
