--- src/feature/community/index.ts
@@ -1,14 +1,0 @@
-export { useCommunityQuery } from './api/queries';
-export type {
-  ComplexExercise,
-  Exercise,
-  ExerciseForCreate,
-  ExerciseType,
-  Workout,
-  WorkoutCategory,
-  WorkoutComment,
-  WorkoutDetail,
-  WorkoutMember,
-} from './model/types';
-export { CommunityPost } from './ui/CommunityPost';
-export { NoPosts } from './ui/NoPosts';
