+++ src/feature/workout/api/queries.ts
@@ -1,9 +1,11 @@
 import { useInfiniteQuery, useQuery } from '@tanstack/react-query';
+import { AxiosError } from 'axios';
 
 import { authApi } from '@/entity/auth';
 import { BaseError, BaseResponse, Pageable } from '@/shared/api';
 
 import {
+  CommunityWorkout,
   ExerciseType,
   Workout,
   WorkoutCategory,
@@ -12,7 +14,8 @@ import {
 } from '../model/types';
 import { WorkoutComment } from '../model/types';
 
-const DEFAULT_SIZE = 20;
+const DEFAULT_WORKOUT_SIZE = 20;
+const DEFAULT_COMMUNITY_SIZE = 10;
 
 export const useWorkoutCategoryListQuery = () => {
   return useQuery<WorkoutCategory[], BaseError>({
@@ -37,7 +40,7 @@ interface WorkoutCommentRequest {
 
 export const useWorkoutCommentQuery = ({
   workoutHistoryId,
-  size = DEFAULT_SIZE,
+  size = DEFAULT_WORKOUT_SIZE,
 }: WorkoutCommentRequest) => {
   return useInfiniteQuery<WorkoutCommentResponse, BaseError>({
     queryKey: ['workoutComment', { workoutHistoryId }],
@@ -83,7 +86,7 @@ interface WorkoutResponse extends Pageable {
 export const useWorkoutQuery = ({
   memberId,
   searchDate,
-  size = DEFAULT_SIZE,
+  size = DEFAULT_WORKOUT_SIZE,
 }: WorkoutRequest) => {
   return useInfiniteQuery<WorkoutResponse, BaseError>({
     queryKey: ['workoutList', { memberId, searchDate }],
@@ -112,7 +115,7 @@ interface WorkoutTypeListResponse extends Pageable {
 
 export const useWorkoutTypeListQuery = ({
   searchValue = '',
-  size = DEFAULT_SIZE,
+  size = DEFAULT_WORKOUT_SIZE,
   exerciseCategory,
 }: WorkoutTypeRequest = {}) => {
   return useInfiniteQuery<WorkoutTypeListResponse, BaseError>({
@@ -139,5 +142,56 @@ export const useWorkoutTypeListQuery = ({
     ) => {
       return !lastPage.isLast ? allPages.length : undefined;
     },
+  });
+};
+
+interface CommunityRequest {
+  searchDate?: string;
+  size?: number;
+  memberId?: string | number | null;
+}
+
+interface CommunityResponse extends Pageable {
+  content: CommunityWorkout[];
+  mainData: WorkoutMember;
+}
+
+export const useCommunityQuery = ({
+  searchDate,
+  size = DEFAULT_COMMUNITY_SIZE,
+  memberId,
+}: CommunityRequest = {}) => {
+  return useInfiniteQuery<CommunityResponse, BaseError>({
+    queryKey: ['workoutList', { memberId }],
+    queryFn: async ({ pageParam }) => {
+      const queryParams = new URLSearchParams();
+      queryParams.append('page', (pageParam as number).toString());
+      queryParams.append('size', size.toString());
+      if (searchDate) {
+        queryParams.append('searchDate', searchDate.toString());
+      }
+      if (memberId) {
+        queryParams.append('memberId', memberId.toString());
+      }
+
+      try {
+        const res = await authApi.get<BaseResponse<CommunityResponse>>(
+          `/api/community/v1?${queryParams.toString()}`
+        );
+        return res.data.data;
+      } catch (error: unknown) {
+        if (error instanceof AxiosError) {
+          const responseData = error.response?.data as { message?: string };
+          const message = responseData.message ?? '문제가 발생했습니다.';
+          throw new Error(message);
+        }
+        throw new Error('문제가 발생했습니다.');
+      }
+    },
+    initialPageParam: 0,
+    getNextPageParam: (lastPage: CommunityResponse, allPages: CommunityResponse[]) => {
+      return !lastPage.isLast ? allPages.length : undefined;
+    },
+    retry: 1,
   });
 };
