+++ src/feature/workout/index.ts
@@ -11,6 +11,7 @@ export {
   useLikeMutation as useWorkoutLikeMutation,
 } from './api/mutations';
 export {
+  useCommunityQuery,
   useWorkoutCategoryListQuery,
   useWorkoutCommentQuery,
   useWorkoutDetailQuery,
@@ -36,6 +37,8 @@ export type {
 export { AppendNewExerciseType } from './ui/AppendNewExerciseType';
 export { CommentInput as WorkoutCommentInput } from './ui/CommentInput';
 export { CommentsWrapper as WorkoutCommentsWrapper } from './ui/CommentList';
+export { CommunityNoPosts } from './ui/CommunityNoPosts';
+export { CommunityPost } from './ui/CommunityPost';
 export { ExerciseDetail, ExercisePreview } from './ui/ExerciseInfo';
 export { NoWorkout } from './ui/NoWorkout';
 export { PostMetrics } from './ui/PostMetrics';
